+++ src/main/java/alexthw/not_enough_glyphs/init/Registry.java
@@ -5,6 +5,7 @@ import alexthw.not_enough_glyphs.common.spell.TrailingProjectile;
 import alexthw.not_enough_glyphs.common.spellbinder.BinderCasterData;
 import alexthw.not_enough_glyphs.common.spellbinder.SpellBinder;
 import alexthw.not_enough_glyphs.common.spellbinder.SpellBinderContainer;
+import com.hollingsworth.arsnouveau.api.perk.PerkAttributes;
 import com.hollingsworth.arsnouveau.common.potions.PublicEffect;
 import com.hollingsworth.arsnouveau.setup.registry.CreativeTabRegistry;
 import net.minecraft.core.component.DataComponentType;
@@ -47,6 +48,7 @@ public class Registry {
 
     public static final DeferredHolder<MobEffect, MobEffect> GROWING_EFFECT = EFFECTS.register("grow", () -> new PublicEffect(MobEffectCategory.NEUTRAL, 0).addAttributeModifier(Attributes.SCALE, prefix("effects.grow"), 0.5D, AttributeModifier.Operation.ADD_VALUE));
     public static final DeferredHolder<MobEffect, MobEffect> SHRINKING_EFFECT = EFFECTS.register("shrink", () -> new PublicEffect(MobEffectCategory.NEUTRAL, 0).addAttributeModifier(Attributes.SCALE, prefix("effects.shrink"), -0.2D, AttributeModifier.Operation.ADD_VALUE));
+    public static final DeferredHolder<MobEffect, MobEffect> STUFFED_EFFECT = EFFECTS.register("stuffed", () -> new PublicEffect(MobEffectCategory.NEUTRAL, 0).addAttributeModifier(Attributes.SCALE, prefix("effects.stuffed"), 0.025D, AttributeModifier.Operation.ADD_VALUE).addAttributeModifier(PerkAttributes.WHIRLIESPRIG, prefix("effects.stuffed"), 0.05, AttributeModifier.Operation.ADD_VALUE));
 
     static {
         TRAILING_PROJECTILE = addEntity("trail", 0.5F, 0.5F, true, true, TrailingProjectile::new, MobCategory.MISC);